Protected game session and update username/password endpoints. Added LastPasswordChange for further validation

+++ SITracker/Controllers/GameSessionsController.cs
@@ -7,6 +7,7 @@ using System.Security.Claims;
 
 namespace SITracker.Controllers
 {
+    [Authorize]
     [ApiController]
     [Route("api/v1/game-sessions")]
     public class GameSessionsController : ControllerBase
@@ -20,7 +21,6 @@ namespace SITracker.Controllers
             _configuration = configuration;
         }
 
-        [Authorize]
         [HttpGet("my-game-sessions")]
         public async Task<ActionResult<List<GameSession>>> GetMyGameSessions()
         {
@@ -33,18 +33,46 @@ namespace SITracker.Controllers
 
             var gameSessions = await _gameSessionService.GetGameSessionsByUserId(long.Parse(userId));
             return Ok(gameSessions);
-
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<GameSession>> GetGameSessionById(long id)
         {
-            return await _gameSessionService.GetGameSessionById(id);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
+            var gameSessionActionResult = await _gameSessionService.GetGameSessionById(id);
+            var gameSession = gameSessionActionResult.Value;
+
+            // Ensure that the user logged in owns the game session found
+            if (gameSession == null)
+            {
+                return NotFound();
+            }
+
+            if (long.Parse(userId) != gameSession.UserId)
+            {
+                return Forbid();
+            }
+
+            return Ok(gameSession);
         }
 
         [HttpPost]
         public async Task<ActionResult<GameSession>> CreateGameSession([FromBody] GameSessionDto createGameSessionDto)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            // Ensure that user that is logged in (jwt) matches the user attached to game session
+            if (userId == null || long.Parse(userId) != createGameSessionDto.UserId)
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var createdGameSessionResult = await _gameSessionService.CreateGameSession(createGameSessionDto);
@@ -67,6 +95,14 @@ namespace SITracker.Controllers
         [HttpPut("{id}")]
         public async Task<ActionResult<GameSession>> UpdateGameSession(long id, [FromBody] GameSessionDto newGameSession)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            // Ensure that user that is logged in (jwt) matches the user attached to game session
+            if (userId == null || long.Parse(userId) != newGameSession.UserId)
+            {
+                return Unauthorized();
+            }
+
             try
             {
                 var updatedGameSession = await _gameSessionService.UpdateGameSession(id, newGameSession);
@@ -84,8 +120,30 @@ namespace SITracker.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteGameSession(long id)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            // Ensure that user is loggedin
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
             try
             {
+                var gameSessionActionResult = await _gameSessionService.GetGameSessionById(id);
+                var gameSession = gameSessionActionResult.Value;
+
+                // Ensure that the user logged in owns the game session found
+                if (gameSession == null)
+                {
+                    return NotFound();
+                }
+
+                if (long.Parse(userId) != gameSession.UserId)
+                {
+                    return Forbid();
+                }
+
                 var result = await _gameSessionService.DeleteGameSession(id);
 
                 if (result == null) return NotFound("Game session not found!");
